+++ Develop/server.js
@@ -1,4 +1,4 @@
-const { mainPrompt } = require('./js')
+const { init } = require('./js/index')
 const cTable = require('console.table')
 
-mainPrompt()
+init()
